moves the data out of home component
