+++ frontend/src/Containers/Home/data.js
@@ -1,0 +1,58 @@
+const ingredientList = ["flour", "sugar", "salt", "butter", "milk"]
+const recipeNames = [
+  "Classic",
+  "Yellow",
+  "Chocolate",
+  "Frosting",
+  "Coconut",
+  "Bundt",
+  "White",
+  "Chocolate-Coconut",
+  "Glaze",
+  "Cream",
+  "Cheese",
+  "Pound",
+  "German",
+  "Chocolate",
+  "Homemade",
+  "Rum",
+  "Peanut",
+  "Butter",
+  "Cup",
+  "Overload",
+  "Salted",
+  "Caramel",
+  "Apple",
+  "Snickers",
+  "Texas",
+  "Sheet",
+  "Ultimate",
+  "Cookies",
+  "Cream",
+  "Oreo",
+  "Cake",
+  "Apple-Cinnamon",
+  "Banana",
+  "Berrylicious",
+  "Blueberry",
+  "Chocolate",
+  "Chip",
+  "Cornbread",
+  "Honey",
+  "Morning",
+  "Glory",
+  "Pumpkin",
+  "Cream",
+  "Cheese",
+  "Pecan",
+  "Streusel",
+  "Triple",
+  "Chocolate",
+  "Chunk",
+  "Zucchini-Chocolate",
+  "Chip",
+  "Muffins",
+  "",
+]
+
+export { ingredientList, recipeNames }
