+++ frontend/src/Containers/Home/index.js
@@ -15,63 +15,7 @@ import ListItem from "@material-ui/core/ListItem"
 import ListItemText from "@material-ui/core/ListItemText"
 import * as actions from "../../actions"
 import Recipe from "../Recipe"
-
-const ingredientList = ["flour", "sugar", "salt", "butter", "milk"]
-const recipeNames = [
-  "Classic",
-  "Yellow",
-  "Chocolate",
-  "Frosting",
-  "Coconut",
-  "Bundt",
-  "White",
-  "Chocolate-Coconut",
-  "Glaze",
-  "Cream",
-  "Cheese",
-  "Pound",
-  "German",
-  "Chocolate",
-  "Homemade",
-  "Rum",
-  "Peanut",
-  "Butter",
-  "Cup",
-  "Overload",
-  "Salted",
-  "Caramel",
-  "Apple",
-  "Snickers",
-  "Texas",
-  "Sheet",
-  "Ultimate",
-  "Cookies",
-  "Cream",
-  "Oreo",
-  "Cake",
-  "Apple-Cinnamon",
-  "Banana",
-  "Berrylicious",
-  "Blueberry",
-  "Chocolate",
-  "Chip",
-  "Cornbread",
-  "Honey",
-  "Morning",
-  "Glory",
-  "Pumpkin",
-  "Cream",
-  "Cheese",
-  "Pecan",
-  "Streusel",
-  "Triple",
-  "Chocolate",
-  "Chunk",
-  "Zucchini-Chocolate",
-  "Chip",
-  "Muffins",
-  "",
-]
+import { ingredientList, recipeNames } from "./data"
 
 class Home extends Component {
   constructor(props) {
